create general TimeBin class
